changed cursor

+++ src/opened_file.h
@@ -74,7 +74,7 @@ public:
     inline int get_num_lines() const { return static_cast<int>(lines.size()); }
     inline int get_num_characters(int line_number = -1) const { 
         if (line_number == -1) line_number = current_line; 
-        return static_cast<int>(lines[line_number].size()) - static_cast<size_t>(line_number == lines.size() - 1); 
+        return static_cast<int>(lines[line_number].size()) - static_cast<size_t>(static_cast<size_t>(line_number) == lines.size() - 1); 
     }
     inline void set_current_line(int line) { current_line = line; }
     inline void set_current_character(int character) { current_character = character; }
